requirements.txt
python-pptx>=0.6.21
pyyaml>=6.0
click>=8.1.0
rich>=13.0
tiktoken>=0.5.0
mcp>=1.0.0